requirements.txt
djangorestframework
django==2.2.3
gunicorn==19.9.0
pyyaml
